pyproject.toml
[project]
name = "aiogramm-tg-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "RasulbekGit",email = "rasuliztileuov@gmail.com"}
]
requires-python = ">=3.14,<3.15"
dependencies = [
    "aiogram (>=3.25.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
